pyproject.toml
[project]
name = "pmi-analyzer"
version = "0.1.0"
description = "PMI Feature Importance Analysis for Manufacturing Process Prediction"
requires-python = ">=3.8"
dependencies = [
    "pandas>=1.3.0",
    "numpy>=1.21.0",
    "scikit-learn>=1.0.0",
    "matplotlib>=3.4.0",
    "seaborn>=0.11.0",
    "pyyaml>=5.4.0",
    "tqdm>=4.62.0",
]

[project.scripts]
pmi-analyze = "pmi_analyzer.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["pmi_analyzer"]
package-dir = {"" = "src"}